Extract napari-workflows-task input models into separate module
