+++ fractal_tasks_core/dev/lib_args_schemas.py
@@ -52,12 +52,12 @@ FRACTAL_TASKS_CORE_PYDANTIC_MODELS = [
     ("fractal_tasks_core", "channels.py", "ChannelInputModel"),
     (
         "fractal_tasks_core",
-        "tasks/napari_workflows_wrapper.py",
+        "tasks/napari_workflows_wrapper_models.py",
         "NapariWorkflowsInput",
     ),
     (
         "fractal_tasks_core",
-        "tasks/napari_workflows_wrapper.py",
+        "tasks/napari_workflows_wrapper_models.py",
         "NapariWorkflowsOutput",
     ),
 ]
