+++ fractal_tasks_core/tasks/napari_workflows_wrapper.py
@@ -15,8 +15,6 @@ Wrapper of napari-workflows.
 import logging
 from pathlib import Path
 from typing import Any
-from typing import Literal
-from typing import Optional
 from typing import Sequence
 
 import anndata as ad
@@ -26,12 +24,9 @@ import numpy as np
 import pandas as pd
 import zarr
 from napari_workflows._io_yaml_v1 import load_workflow
-from pydantic import BaseModel
-from pydantic import validator
 from pydantic.decorator import validate_arguments
 
 import fractal_tasks_core
-from fractal_tasks_core.channels import ChannelInputModel
 from fractal_tasks_core.channels import get_channel_from_image_zarr
 from fractal_tasks_core.labels import prepare_label_group
 from fractal_tasks_core.ngff import load_NgffImageMeta
@@ -42,6 +37,12 @@ from fractal_tasks_core.roi import (
 )
 from fractal_tasks_core.roi import load_region
 from fractal_tasks_core.tables import write_table
+from fractal_tasks_core.tasks.napari_workflows_wrapper_models import (
+    NapariWorkflowsInput,
+)
+from fractal_tasks_core.tasks.napari_workflows_wrapper_models import (
+    NapariWorkflowsOutput,
+)
 from fractal_tasks_core.upscale_array import upscale_array
 from fractal_tasks_core.utils import rescale_datasets
 
@@ -58,72 +59,6 @@ class OutOfTaskScopeError(NotImplementedError):
     """
 
     pass
-
-
-class NapariWorkflowsInput(BaseModel):
-    """
-    A value of the `input_specs` argument in `napari_workflows_wrapper`.
-
-    Attributes:
-        type: Input type (either `image` or `label`).
-        label_name: Label name (for label inputs only).
-        channel: `ChannelInputModel` object (for image inputs only).
-    """
-
-    type: Literal["image", "label"]
-    label_name: Optional[str]
-    channel: Optional[ChannelInputModel]
-
-    @validator("label_name", always=True)
-    def label_name_is_present(cls, v, values):
-        """
-        Check that label inputs have `label_name` set.
-        """
-        _type = values.get("type")
-        if _type == "label" and not v:
-            raise ValueError(
-                f"Input item has type={_type} but label_name={v}."
-            )
-        return v
-
-    @validator("channel", always=True)
-    def channel_is_present(cls, v, values):
-        """
-        Check that image inputs have `channel` set.
-        """
-        _type = values.get("type")
-        if _type == "image" and not v:
-            raise ValueError(f"Input item has type={_type} but channel={v}.")
-        return v
-
-
-class NapariWorkflowsOutput(BaseModel):
-    """
-    A value of the `output_specs` argument in `napari_workflows_wrapper`.
-
-    Attributes:
-        type: Output type (either `label` or `dataframe`).
-        label_name: Label name (for label outputs, it is used as the name of
-            the label; for dataframe outputs, it is used to fill the
-            `region["path"]` field).
-        table_name: Table name (for dataframe outputs only).
-    """
-
-    type: Literal["label", "dataframe"]
-    label_name: str
-    table_name: Optional[str] = None
-
-    @validator("table_name", always=True)
-    def table_name_only_for_dataframe_type(cls, v, values):
-        """
-        Check that table_name is set only for dataframe outputs.
-        """
-        _type = values.get("type")
-        if (_type == "dataframe" and (not v)) or (_type != "dataframe" and v):
-            raise ValueError(
-                f"Output item has type={_type} but table_name={v}."
-            )
-        return v
 
 
 @validate_arguments
