+++ tests/tasks/test_workflows_napari_workflows.py
@@ -28,10 +28,10 @@ from .lib_empty_ROI_table import _add_empty_ROI_table
 from fractal_tasks_core.tasks.napari_workflows_wrapper import (
     napari_workflows_wrapper,
 )
-from fractal_tasks_core.tasks.napari_workflows_wrapper import (
+from fractal_tasks_core.tasks.napari_workflows_wrapper_models import (
     NapariWorkflowsInput,
 )
-from fractal_tasks_core.tasks.napari_workflows_wrapper import (
+from fractal_tasks_core.tasks.napari_workflows_wrapper_models import (
     NapariWorkflowsOutput,
 )
 from fractal_tasks_core.zarr_utils import OverwriteNotAllowedError
